component CircleInfoCard created, so now it is available in Home and Proposals Page
