+++ src/components/Cards/ServerInfoCard.jsx
@@ -14,9 +14,9 @@ const ServerInfoCard = () => {
       title="Free Server Notice"
       description="This is a small project running on free servers.
       
-       So login and registration might take a moment longer than usual.
+       Sometimes login and registration might take a moment longer than usual.
        
-       Once you're in, the servers are warmed up and running smoothly!"
+       But once you're in, the servers are warmed up and running smoothly!"
       className="bg-blue-50 border-blue-200 p-4"
       titleClassName="text-blue-300"
       descriptionClassName="text-blue-400 text-justify whitespace-pre-line"
