+++ src/components/Cards/index.js
@@ -8,3 +8,4 @@ export { default as MatchMemberCard } from "./MatchMemberCard.jsx";
 export { default as ProposalCard } from "./ProposalCard.jsx";
 export { default as GenericCard } from "./GenericCard.jsx";
 export { default as ServerInfoCard } from "./ServerInfoCard.jsx";
+export { default as CircleInfoCard } from "./CircleInfoCard.jsx";
